Added statically generated version of site, moved search from sidebar
The static site should help search engines make more sense of things.
Maybe. Who knows.

The long titles of many of the picks was causing the sidebar search to
be way too big. I had set the sidebar to expand with the window, but
that tended to not look great, and was even worse in the static site
because nothing was even there. Instead, I made a search page similar to
the advanced search.