Small tweaks to search
Finally fixed the missing border around the text box by just plopping
some CSS on it, there's probably something more fundamental to be done
to fix it but meh.

Also removed the reveal widget that prevented a blank search box from
showing any results. That's good default behavior but not how I want it
to work. There might be more to clean up in there since I just
copy-pasted most of that, but for now it seems to do what I want.