New topics, advanced search improvements
The new topics: Live Edited, Alternate Risky, Passionate, and No
Passion.

The advanced search now has an "any/all" toggle for the topic section.
The other search criteria don't really need one, since an AND operation
on them will always be empty. I also changed the temp search tiddlers to
be marked as "volatile" which will help the text box to feel less
sluggish when the advanced search tab is open. For volatile tiddlers,
the system waits to re-render until the tiddler has settled down for a
specified period of time.